[clean] Create a new PendingChat component
